+++ src/types/MatrixDense/MatrixDense_hlf.cu
@@ -240,10 +240,9 @@ MatrixDense<float> MatrixDense<__half>::to_float() const {
     
     MatrixDense<float> created_mat(cu_handles, m_rows, n_cols);
 
-    double NUM_THREADS = 1024; // threads per thread block just 1 warp
-    double NUM_BLOCKS = static_cast<double>(
-        std::ceil(static_cast<double>(m_rows*n_cols)/static_cast<double>(NUM_THREADS))
-    );
+    int NUM_THREADS = 1024; // threads per thread block just maximum
+    int NUM_BLOCKS = std::ceil(static_cast<double>(m_rows*n_cols)/static_cast<double>(NUM_THREADS));
+
     generalmatrix_hlf_kernels::cast_to_float<<<NUM_THREADS, NUM_BLOCKS>>>(
         d_mat, created_mat.d_mat, m_rows*n_cols
     );
@@ -256,10 +255,9 @@ MatrixDense<double> MatrixDense<__half>::to_double() const {
     
     MatrixDense<double> created_mat(cu_handles, m_rows, n_cols);
 
-    double NUM_THREADS = 1024; // threads per thread block just 1 warp
-    double NUM_BLOCKS = static_cast<double>(
-        std::ceil(static_cast<double>(m_rows*n_cols)/static_cast<double>(NUM_THREADS))
-    );
+    int NUM_THREADS = 1024; // threads per thread block just maximum
+    int NUM_BLOCKS = std::ceil(static_cast<double>(m_rows*n_cols)/static_cast<double>(NUM_THREADS));
+
     generalmatrix_hlf_kernels::cast_to_double<<<NUM_THREADS, NUM_BLOCKS>>>(
         d_mat, created_mat.d_mat, m_rows*n_cols
     );
